no message
svn path=/trunk/yarp2/; revision=6723

+++ src/modules/fakebot/FakeBot.cpp
@@ -146,7 +146,7 @@ bool FakeBot::getImage(yarp::sig::ImageOf<yarp::sig::PixelRgb>& image) {
     }
     IMGFOR(image,x2,y2) {
         PixelRgb& pix = image(x2,y2);
-        float f = noiseLevel;
+        float f = (float)(noiseLevel);
         scramble(pix.r,f);
         scramble(pix.g,f);
         scramble(pix.b,f);
